pref: add pre-generated static paths
