+++ src/lib/content/paths.ts
@@ -2,10 +2,9 @@ import { memoize } from 'lodash';
 import path from 'path';
 import {
   getAllMarkdownFiles,
-  getRedirectsNotToAliases,
-  getReversedAliasPaths,
   normalizePathWithSlash,
 } from '../../../scripts/common';
+import { readFileSync } from 'fs';
 
 export { getAllMarkdownFiles };
 
@@ -13,6 +12,8 @@ export function getContentPath(slug: string) {
   const contentDir = path.join(process.cwd(), 'public/content');
   return path.join(contentDir, slug);
 }
+
+let CACHED_JSON_PATH: Record<string, string> | null = null;
 
 /**
  * Generates a map of static JSON paths for the application.
@@ -30,64 +31,24 @@ export function getContentPath(slug: string) {
  * @returns A promise that resolves to a Record mapping browser path strings to their target Markdown path strings,
  *          where all paths are normalized with slashes.
  */
-export const getStaticJSONPaths = memoize(
-  async (): Promise<Record<string, string>> => {
-    const contentDir = path.join(process.cwd(), 'public/content');
-    const redirectsPaths = await getRedirectsNotToAliases();
-    const aliasesPaths = await getReversedAliasPaths();
-    const markdownPaths = (await getAllMarkdownFiles(contentDir))
-      .filter(
-        slugArray =>
-          !slugArray[0]?.toLowerCase()?.startsWith('contributor') &&
-          !slugArray[0]?.toLowerCase()?.startsWith('tags'),
-      )
-      .map(slugArray => slugArray.join('/'));
+export const getStaticJSONPaths = async (): Promise<Record<string, string>> => {
+  if (CACHED_JSON_PATH) {
+    return CACHED_JSON_PATH;
+  }
+  const staticJsonPath = path.join(
+    process.cwd(),
+    'public/content/static-paths.json',
+  );
 
-    // Aliases paths as primary paths
-    const aliasesEntries = Object.entries(aliasesPaths);
-
-    // Only getting redirects paths that are unique and not already in aliases
-    const redirectsPathsToAliases = Object.entries(redirectsPaths);
-
-    // Removing all markdown paths that are already in redirects and aliases
-    const filteredMarkdownPaths = markdownPaths
-      .filter(mdPath => {
-        const normalizedPath = normalizePathWithSlash(mdPath);
-        const isMatchedRedirects = redirectsPathsToAliases.some(
-          ([redirectKey]) => {
-            const normalizedRedirectKey = normalizePathWithSlash(redirectKey);
-            return normalizedPath === normalizedRedirectKey;
-          },
-        );
-        if (isMatchedRedirects) {
-          return false;
-        }
-
-        const isMatchedAliases = aliasesEntries.some(([aliasKey]) => {
-          const normalizedAliasKey = normalizePathWithSlash(aliasKey);
-          return normalizedPath === normalizedAliasKey;
-        });
-        return !isMatchedAliases;
-      })
-      .map(mdPath => [mdPath, mdPath]);
-    const aliasesEntriesPaths = aliasesEntries.map(([key, value]) => {
-      return [value, key];
-    });
-
-    const paths: Record<string, string> = {
-      ...Object.fromEntries(filteredMarkdownPaths),
-      ...redirectsPaths,
-      ...Object.fromEntries(aliasesEntriesPaths),
-    };
-
-    return Object.fromEntries(
-      Object.entries(paths).map(([key, value]) => [
-        normalizePathWithSlash(key),
-        normalizePathWithSlash(value),
-      ]),
-    );
-  },
-);
+  try {
+    const jsonData = readFileSync(staticJsonPath, 'utf-8');
+    CACHED_JSON_PATH = JSON.parse(jsonData) as Record<string, string>;
+    return CACHED_JSON_PATH ?? {};
+  } catch (error) {
+    console.error('Error reading static paths JSON:', error);
+    return {};
+  }
+};
 
 export const getRedirectsBackLinks = memoize(
   (
